astro-app/src/requirements.txt
Django
djangorestframework
psycopg2-binary
google-generativeai
python-dotenv
django-cors-headers